requirements.txt
flask==2.3.3
flask-login==0.6.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
pgvector==0.2.4
langchain==0.1.0
langchain-core==0.1.23
langchain-text-splitters==0.0.1
langchain-google-genai==0.0.11
google-generativeai==0.8.5
werkzeug==2.3.7
python-dotenv==1.0.0
pypdf2==3.0.1
python-docx==0.8.11
pandas==2.1.0
numpy==1.24.3
gunicorn==21.2.0
